+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to this project will be documented in this file.
 - Stylus all pages & components (maxwidth = $breakpoint-xs quasar's variable)
 - Admin Default layout & Login
 - Customer Detail Page with Sign-Out feaature
+- Validator Login form[@phillip](https://github.com/hemvanh/TPTEA/pull/23)
+- Routing with new token of Admin & Customer[@phillip](https://github.com/hemvanh/TPTEA/pull/24)
+- Test case routing (No-Token, TokenAdmin, TokenCustomer)
+- Automation Test Routing (Testcafe framework)
 - Schema Admin (res loginAdmin,def type Admin,input AdminLoginInput)[@john]
 ### Removed 
 - Join now button